mrc-kv-cache-configurator-core(kvc-move-dsv4-compress-dtypes,non_mechanical_provable): Move _get_dsv4_compress_state_dtypes to kv_cache_configurator (cut+paste)
