+++ python/sglang/srt/model_executor/model_runner_kv_cache_mixin.py
@@ -86,23 +86,6 @@ MAMBA_CACHE_V2_ADDITIONAL_RATIO_NO_OVERLAP = 1
 logger = logging.getLogger(__name__)
 
 
-def _get_dsv4_compress_state_dtypes() -> tuple[torch.dtype, torch.dtype]:
-    dtype_name = envs.SGLANG_DSV4_COMPRESS_STATE_DTYPE.get().strip().lower()
-    if dtype_name in ("float32", "fp32"):
-        return torch.float32, torch.float32
-    if dtype_name in ("bfloat16", "bf16"):
-        if envs.SGLANG_OPT_USE_ONLINE_COMPRESS.get():
-            raise ValueError(
-                "SGLANG_DSV4_COMPRESS_STATE_DTYPE=bf16 is not supported when "
-                "SGLANG_OPT_USE_ONLINE_COMPRESS=1; online c128 state must stay float32."
-            )
-        return torch.bfloat16, torch.bfloat16
-    raise ValueError(
-        "Unsupported SGLANG_DSV4_COMPRESS_STATE_DTYPE="
-        f"{dtype_name!r}. Expected one of: float32, fp32, bfloat16, bf16."
-    )
-
-
 _is_npu = is_npu()
 _is_hip = is_hip()
 
@@ -1382,6 +1365,10 @@ class ModelRunnerKVCacheMixin:
         # Draft worker does not own the compression-state pools, but keep the
         # dtype attributes initialized so _init_pools can share one code path.
         if is_deepseek_v4(self.model_config.hf_config):
+            from sglang.srt.mem_cache.kv_cache_configurator import (
+                _get_dsv4_compress_state_dtypes,
+            )
+
             self.c4_state_dtype, self.c128_state_dtype = (
                 _get_dsv4_compress_state_dtypes()
             )
